% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_init.R
\name{colorZapper_file_active}
\alias{colorZapper_file_active}
\title{Checks for an active colorZapper project.
Checks for an active colorZapper project.}
\usage{
colorZapper_file_active()
}
\description{
Checks for an active colorZapper project.
Checks for an active colorZapper project.
}
