% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_interact.R
\name{CZcheck}
\alias{CZcheck}
\title{Check out defined ROI-s
Check out defined ROI-s}
\usage{
CZcheck(file = tempfile(fileext = ".pdf"))
}
\description{
Check out defined ROI-s
Check out defined ROI-s
}
\examples{
\dontrun{
require(colorZapper)
dir = system.file(package = "colorZapper", "sample")
CZopen(path = tempfile() )
CZaddFiles(dir)
CZdefine(points = 1)
CZdefine(polygons = 1, what = 1)
CZcheck()
}
}
