% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_process.R
\name{CZextractALL}
\alias{CZextractALL}
\title{Extract data
Extract data from the entire image}
\usage{
CZextractALL(parralel = TRUE)
}
\description{
Extract data
Extract data from the entire image
}
\examples{
\dontrun{
 CZextractALL()
 }
}
