% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_init.R
\name{CZopen}
\alias{CZopen}
\title{Open a colorZapper file.
Setup a new colorZapper project or open an existing file.}
\usage{
CZopen(path, overwrite = FALSE)
}
\arguments{
\item{path}{file path.}

\item{overwrite}{when TRUE, start from scratch; default to FALSE.}
}
\description{
Open a colorZapper file.
Setup a new colorZapper project or open an existing file.
}
\examples{
\dontrun{
require(colorZapper)
dir = system.file(package = "colorZapper", "sample")
CZopen(path = tempfile() )

}
}
