% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_init.R
\name{CZopen_example}
\alias{CZopen_example}
\title{Open the example colorZapper file.
The file open by CZopen_example() has the interactive ROI-s defined; it is used for testing the non-interactive functions.}
\usage{
CZopen_example()
}
\description{
Open the example colorZapper file.
The file open by CZopen_example() has the interactive ROI-s defined; it is used for testing the non-interactive functions.
}
\examples{
require(colorZapper)
CZopen_example()
}
