% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_process.R
\name{CZrm}
\alias{CZrm}
\title{Remove entries
Remove entries associated with given images}
\usage{
CZrm(fileNam)
}
\arguments{
\item{fileNam}{File name whose corresponding entries will be removed from the active project.}
}
\description{
Remove entries
Remove entries associated with given images
}
