% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_get_data.R
\name{CZshowStatus}
\alias{CZshowStatus}
\title{Show db status
Returns a data.frame containing the status of the current CZ project.}
\usage{
CZshowStatus()
}
\description{
Show db status
Returns a data.frame containing the status of the current CZ project.
}
\examples{
CZopen_example()
CZshowStatus()
}
