% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_data.R
\name{db2list}
\alias{db2list}
\title{db as a list of data.frame-s
db as a list of data.frame-s}
\usage{
db2list(con)
}
\description{
db as a list of data.frame-s
db as a list of data.frame-s
}
\examples{
\dontrun{
 db2list(con)
 }
}
