% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_birds.R
\name{generateBirds}
\alias{generateBirds}
\title{Title generate birds
Title generate n species (both sexes) with males (slightly) more colorful than females}
\usage{
generateBirds(n, dir = tempdir(), body_parts, res = 10)
}
\arguments{
\item{n}{No of species}

\item{dir}{Directory to save images}

\item{body_parts}{one of: "crown", "forehead", "lore", "chin", "throat", "nape", "ear-coverts", "breast", "mantle",  "scapulars", "coverts", "alula", "flight-feathers", "flank", "back", "rump", "belly", "vent", "tail".  Default to all}

\item{res}{Resolution, in dots per inch.}
}
\value{
dir
}
\description{
Title generate birds
Title generate n species (both sexes) with males (slightly) more colorful than females
}
\examples{
f = generateBirds(n = 10) 
f = generateBirds(n = 10, body_parts = "flight-feathers") 

}
