% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloration_score.R
\name{color_scores}
\alias{color_scores}
\title{Compute color scores}
\usage{
color_scores(d, rgb = c("R", "G", "B"), id = "id", sex = "sex",
  maleLevel = "m.png", group = mean, prob = 0.01, ...)
}
\description{
Compute color scores
}
\examples{
/donotrun{

f = generateBirds(n = 10) 
cz_file = tempfile(fileext = '.sqlite')
CZopen(path = cz_file)
CZaddFiles(f)
CZextractALL()
d = CZdata(what = 'ALL')
d[, sex := stringr::str_split(path, '_', simplify = TRUE)[2], by = 1:nrow(d)]

x =color_scores(d)
}

}
