% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{colorZapper}
\alias{.onLoad}
\alias{colorZapper}
\alias{colorZapper-package}
\title{Color zapper.}
\usage{
.onLoad(libname, pkgname)
}
\description{
Color zapper.
}
\section{Package options}{


colorZapper description:

colorZapper session:

\itemize{
\item \code{zzz}: path to use for \code{\link{zzz}}
}
}

