% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AAA.R
\docType{package}
\name{colorZapper}
\alias{colorZapper}
\title{colorZapper.}
\description{
Colour extraction utilities.
}
\author{
Mihai Valcu \email{valcu@orn.mpg.de},
James Dale \email{J.Dale@massey.ac.nz}

Maintainer: Mihai Valcu \email{valcu@orn.mpg.de}
}
