% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CZAddFiles}
\alias{CZAddFiles}
\title{Associate images to a colorZapper file.}
\usage{
CZAddFiles(dir)
}
\arguments{
\item{dir}{directory containing images}
}
\value{
TRUE
}
\description{
Associate images to a colorZapper file.
}
\note{
duplicated files are dropped silently
}
\examples{
\dontrun{
require(colorZapper)
dir = system.file(package = "colorZapper", "sample")
CZOpen(path = tempfile() )
CZAddFiles(dir)

}
}

