% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_setup.R
\name{CZaddFiles}
\alias{CZaddFiles}
\alias{CZsetwd}
\title{Associate images to a colorZapper file.}
\usage{
CZaddFiles(dir)

CZsetwd(dir = getwd())
}
\arguments{
\item{dir}{directory containing images. When re-setting dir with CZsetwd() the default is the current working directory.}
}
\value{
TRUE
}
\description{
Associate images to a colorZapper file.
}
\note{
CZaddFiles() can be run only once.
}
\examples{
require(colorZapper)
dir = system.file(package = "colorZapper", "sample")
p = tempfile() 
CZopen(path = p)
CZaddFiles(dir)
CZsetwd()
}
