% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CZaddFiles}
\alias{CZaddFiles}
\title{Associate images to a colorZapper file.}
\usage{
CZaddFiles(dir)
}
\arguments{
\item{dir}{directory containing images}
}
\value{
TRUE
}
\description{
Associate images to a colorZapper file.
}
\note{
Duplicated files are dropped silently.
}
\examples{
\dontrun{
require(colorZapper)
dir = system.file(package = "colorZapper", "sample")
CZopen(path = tempfile() )
CZaddFiles(dir)

}
}

