% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_interact.R
\name{CZcheck}
\alias{CZcheck}
\title{Check out defined ROI-s
Plot the defined ROI-s on the original image.}
\usage{
CZcheck(file = tempfile(fileext = ".pdf"))
}
\description{
Check out defined ROI-s
Plot the defined ROI-s on the original image.
}
\note{
Warning: this can take a long time for many images.
}
\examples{
\dontrun{
require(colorZapper)
dir = system.file(package = "colorZapper", "sample")
CZopen(path = tempfile() )
CZaddFiles(dir)
CZdefine(points = 1)
CZdefine(polygons = 1, what = 1)
CZcheck()
}
}
