% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_get_data.R
\name{CZdata}
\alias{CZdata}
\title{colorZapper data
Fetch colorZapper RGB data}
\usage{
CZdata(what)
}
\arguments{
\item{what}{'ROI' (gets the data of ROI-s defined interactively) 
or 'ALL' (extracts the color of from all images.)}
}
\description{
colorZapper data
Fetch colorZapper RGB data
}
\examples{
require(doParallel)
registerDoParallel(1)
CZopen_example()
CZextractROI()
CZextractALL()
stopImplicitCluster()
d = CZdata(what= 'ROI')
d = CZdata(what= 'ALL')

}
