% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CZDefine}
\alias{CZDefine}
\title{Define regions of interest.
Interactively define points or polygons using mouse clicks.}
\usage{
CZDefine(points, polygons, ...)
}
\description{
Define regions of interest.
Interactively define points or polygons using mouse clicks.
}
\examples{
\dontrun{
require(colorZapper)
dir = system.file(package = "colorZapper", "sample")
CZOpen(path = tempfile() )
CZAddFiles(dir)
CZDefine(points = 1)

}
}

