% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CZdefine}
\alias{CZdefine}
\title{Define regions of interest.
Interactively define points or polygons using mouse clicks.}
\usage{
CZdefine(points, polygons, ...)
}
\description{
Define regions of interest.
Interactively define points or polygons using mouse clicks.
}
\examples{
\dontrun{
require(colorZapper)
dir = system.file(package = "colorZapper", "sample")
CZopen(path = tempfile() )
CZaddFiles(dir)
CZdefine(points = 1)

}
}

