% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_process.R
\name{CZextractROI}
\alias{CZextractROI}
\title{Extract data
Extract data from regions of interest.}
\usage{
CZextractROI(parallel = FALSE)
}
\description{
Extract data
Extract data from regions of interest.
}
\examples{
\dontrun{
 CZextractROI()
 }
}
