% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CZopen}
\alias{CZopen}
\title{Open a colorZapper file.}
\usage{
CZopen(path)
}
\description{
Open a colorZapper file.
}
\examples{
\dontrun{
require(colorZapper)
dir = system.file(package = "colorZapper", "sample")
CZopen(path = tempfile() )

}
}

