% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_init.R
\name{CZopen}
\alias{CZopen}
\title{Open a colorZapper file.
Open a colorZapper file.}
\usage{
CZopen(path)
}
\description{
Open a colorZapper file.
Open a colorZapper file.
}
\examples{
\dontrun{
require(colorZapper)
dir = system.file(package = "colorZapper", "sample")
CZopen(path = tempfile() )

}
}
