% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloration_score.R
\name{generateBirds}
\alias{generateBirds}
\title{Title generate birds
Title generate n species (both sexes) with males (slightly) more colorful than females}
\usage{
generateBirds(n, dir = tempdir(), body_parts)
}
\arguments{
\item{n}{No of species}

\item{dir}{Directory to save images}

\item{body_parts}{one of: "crown", "forehead", "lore", "chin", "throat", "nape", "ear-coverts", "breast", "mantle", "scapulars", "coverts", "alula", "flight-feathers", "flank", "back", "rump", "belly", "vent", "tail". Default to all}
}
\value{
dir
}
\description{
Title generate birds
Title generate n species (both sexes) with males (slightly) more colorful than females
}
\examples{
\dontrun{
f = generateBirds(n = 10) 
cz_file = tempfile(fileext = '.sqlite')
CZopen(path = cz_file)
CZaddFiles(f)
CZextractALL()
d = CZdata(what = 'ALL')





} 
}
