% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zap}
\alias{zap}
\title{simple color extraction in a loop, invoking locator aka \code{\link{click}}.}
\usage{
zap(path = system.file(package = "colorZapper", "sample"), N = 3)
}
\arguments{
\item{N}{number of points per image to extract}

\item{string}{directory containing images}
}
\value{
\code{data.frame} containing the file path and the RGB
}
\description{
simple color extraction in a loop, invoking locator aka \code{\link{click}}.
}
\examples{
\dontrun{
zap()


}
}

